try and delete lab1, rename lab4 to lab1, maybe this will work
