+++ src/main/kotlin/lab1/ApplicationLoop.kt
@@ -1,10 +1,10 @@
-package lab4
+package lab1
 
-import lab4.data.FileManager
-import lab4.logic.TaskLogic
-import lab4.model.Task
-import lab4.model.TaskStatus
-import lab4.presentation.ConsoleUI
+import lab1.data.FileManager
+import lab1.logic.TaskLogic
+import lab1.model.Task
+import lab1.model.TaskStatus
+import lab1.presentation.ConsoleUI
 
 
 /**
