+++ src/main/kotlin/lab1/Lab4-callgraph-after-refactor.png
