+++ src/main/kotlin/lab1/presentation/ConsoleUI.kt
@@ -1,4 +1,4 @@
-package lab4.presentation
+package lab1.presentation
 
 
 /**
